backend/requirements.txt
# DeepEcho backend - FastAPI + MentalBERT (local)
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database (Supabase PostgreSQL via SQLAlchemy)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Auth (Supabase Auth)
supabase>=2.0.0
postgrest>=0.13.0
email-validator>=2.0.0

# NLP & model (run locally)
transformers>=4.40.0
torch>=2.0.0
tokenizers>=0.19.0

# Preprocessing
spacy>=3.7.0
# Run after install: python -m spacy download en_core_web_sm

# Hugging Face hub (for one-time download to local folder)
huggingface-hub>=0.26.0

# Env
python-dotenv>=1.0.0
